% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ia2rms.R
\name{evalH}
\alias{evalH}
\title{Title}
\usage{
evalH(x, parLI, parLS, S, intervalos, a)
}
\arguments{
\item{x}{.}

\item{parLI}{.}

\item{parLS}{.}

\item{S}{.}

\item{intervalos}{.}

\item{a}{.}
}
\value{
.
}
\description{
Title
}
\examples{
.
}
