% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ia2rms.R
\name{Ia2rms}
\alias{Ia2rms}
\title{Independent Doubly Adaptive Rejection Metropolis Sampling}
\usage{
Ia2rms(y.initial, logf, indFunc, n.sample)
}
\arguments{
\item{y.initial}{Value inside the domain of the target density}

\item{logf}{log of the target density. (Class function)}

\item{indFunc}{Indicator function for the domain of the target density. (Class function)}

\item{n.sample}{Number of samples to draw from the taret density}
}
\value{
Sample of lenght n.sample
}
\description{
Independent Doubly Adaptive Rejection Metropolis Sampling
}
\examples{
y.initial <- 0.8
n.sample <- 5000
logf <- function(x) log(0.3 * dt(x, 3.2) + 0.35*dnorm(x,mean = 2.5, 0.3) + 0.5*dt(x, 1))
indFunc <- function(x) all(x > -5) * all(x < 6)
y <- Ia2rms(y.initial, logf, indFunc, n.sample)
hist(y[[1]],breaks = 50, probability = T, ylim = c(0,0.5))
curve(exp(Vectorize(logf)(x)), -5, 8, ylab = "f(x)", main = "Representación de f(x) y g(x)", add = T)
legend(legend = c("f(x)", "g(x)"), col = c("black", "blue"), x = "topleft", lty = c(1, 1))

}
