% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ia2rms.R
\name{parLI}
\alias{parLI}
\title{Title}
\usage{
parLI(S, x, sim, eval, cum, logf)
}
\arguments{
\item{S}{.}

\item{x}{.}

\item{sim}{.}

\item{eval}{.}

\item{cum}{.}

\item{logf}{log}
}
\value{
.
}
\description{
Title
}
