% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ia2rms.R
\name{Ia2rms}
\alias{Ia2rms}
\title{Independent Doubly Adaptive Rejection Metropolis Sampling}
\usage{
Ia2rms(y.initial, logf, indFunc, n.sample)
}
\arguments{
\item{y.initial}{Value inside the domain of the target density}

\item{logf}{log of the target density. (Class function)}

\item{indFunc}{Indicator function for the domain of the target density. (Class function)}

\item{n.sample}{Number of samples to draw from the target density}
}
\value{
Sample of lenght n.sample
}
\description{
Independent Doubly Adaptive Rejection Metropolis Sampling
}
\examples{

logf <- function(x) log(prod(0.3,dt(x, 3.2)) + prod(0.35,dnorm(x, 2.5, 0.3)) + prod(0.5,dt(x, 1)))
indFunc <- function(x) all(x > -5) * all(x < 6)
y <- Ia2rms(y.initial = 0.8, logf = logf, indFunc = indFunc, n.sample = 1000)
hist(y,breaks = 50, probability = TRUE, ylim = c(0,0.5))
curve(exp(Vectorize(logf)(x)), -5, 8, ylab = "f(x)", add = TRUE)
legend(legend = c("f(x)", "g(x)"), col = c("black", "blue"), x = "topleft", lty = c(1, 1))

}
