% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ia2rms.R
\name{parLS}
\alias{parLS}
\title{Title.}
\usage{
parLS(S, x, eval, sim, cum)
}
\arguments{
\item{S}{.}

\item{x}{.}

\item{eval}{.}

\item{sim}{.}

\item{cum}{.}
}
\value{
.
}
\description{
Title.
}
\examples{
.
}
